+++ src/commands/visitAtPointCommands.ts
@@ -7,10 +7,13 @@ import { gitRun } from '../utils/gitRawRunner';
 import { CommitDetailView } from '../views/commitDetailView';
 import { views } from '../extension';
 import { StashItemView } from '../views/stashes/stashSectionView';
-import { StashDetailView } from '../views/stashDetailView';
 import { ChangeView } from '../views/changes/changeView';
 import { MagitCommit } from '../models/magitCommit';
 import { HunkView } from '../views/changes/hunkView';
+import { BranchListingView } from '../views/branches/branchListingView';
+import { RemoteBranchListingView } from '../views/remotes/remoteBranchListingView';
+import { TagListingView } from '../views/tags/tagListingView';
+import { showStashDetail } from './diffingCommands';
 
 export async function magitVisitAtPoint(repository: MagitRepository, currentView: DocumentView) {
 
@@ -29,23 +32,28 @@ export async function magitVisitAtPoint(repository: MagitRepository, currentView
   } else if (selectedView instanceof CommitItemView) {
 
     const commit: MagitCommit = (selectedView as CommitItemView).commit;
+    return visitCommit(repository, commit.hash);
 
-    const result = await gitRun(repository, ['show', commit.hash]);
-    commit.diff = result.stdout;
+  } else if (selectedView instanceof BranchListingView ||
+    selectedView instanceof RemoteBranchListingView ||
+    selectedView instanceof TagListingView) {
 
-    const uri = CommitDetailView.encodeLocation(commit.hash);
-    views.set(uri.toString(), new CommitDetailView(uri, commit));
-    workspace.openTextDocument(uri).then(doc => window.showTextDocument(doc, MagitUtils.oppositeActiveViewColumn()));
+    const commit = (selectedView as BranchListingView).ref.commit;
+    return visitCommit(repository, commit!);
 
   } else if (selectedView instanceof StashItemView) {
 
     const stash = (selectedView as StashItemView).stash;
-    const uri = StashDetailView.encodeLocation(repository, stash);
-
-    const result = await gitRun(repository, ['stash', 'show', '-p', `stash@{${stash.index}}`]);
-    const stashDiff = result.stdout;
-
-    views.set(uri.toString(), new StashDetailView(uri, stash, stashDiff));
-    workspace.openTextDocument(uri).then(doc => window.showTextDocument(doc, MagitUtils.oppositeActiveViewColumn()));
+    showStashDetail(repository, stash);
   }
+}
+
+export async function visitCommit(repository: MagitRepository, commitHash: string) {
+  const commit: MagitCommit = { hash: commitHash, message: '', parents: [] };
+  const result = await gitRun(repository, ['show', commitHash]);
+  commit.diff = result.stdout;
+
+  const uri = CommitDetailView.encodeLocation(commit.hash);
+  views.set(uri.toString(), new CommitDetailView(uri, commit));
+  workspace.openTextDocument(uri).then(doc => window.showTextDocument(doc, MagitUtils.oppositeActiveViewColumn()));
 }