+++ src/views/blameView.ts
@@ -3,6 +3,7 @@ import { Uri } from 'vscode';
 import * as Constants from '../common/constants';
 import { TextView } from './general/textView';
 import { MagitRepository } from '../models/magitRepository';
+import { MagitState } from '../models/magitState';
 
 export class BlameView extends DocumentView {
 
@@ -18,7 +19,7 @@ export class BlameView extends DocumentView {
     this.addSubview(blameTextView);
   }
 
-  public update(repository: MagitRepository): void { }
+  public update(state: MagitState): void { }
 
   static encodeLocation(repository: MagitRepository, fileUri: Uri): Uri {
     return Uri.parse(`${Constants.MagitUriScheme}:${BlameView.UriPath}?${repository.rootUri.fsPath}#${fileUri.path}`);