+++ src/views/diffView.ts
@@ -2,11 +2,12 @@ import { DocumentView } from './general/documentView';
 import { Uri } from 'vscode';
 import * as Constants from '../common/constants';
 import { TextView } from './general/textView';
+import { MagitState } from '../models/magitState';
 import { MagitRepository } from '../models/magitRepository';
 
 export class DiffView extends DocumentView {
 
-  static UriPath: string = 'magit.diff';
+  static UriPath: string = 'diff.magit';
   isHighlightable = false;
   needsUpdate = false;
 
@@ -18,10 +19,10 @@ export class DiffView extends DocumentView {
     this.addSubview(diffTextView);
   }
 
-  public update(repository: MagitRepository): void { }
+  public update(state: MagitState): void { }
 
   static index = 0;
-  static encodeLocation(diffId: string): Uri {
-    return Uri.parse(`${Constants.MagitUriScheme}:${DiffView.UriPath}?${diffId}#${DiffView.index++}`);
+  static encodeLocation(repository: MagitRepository, diffId: string): Uri {
+    return Uri.parse(`${Constants.MagitUriScheme}:${DiffView.UriPath}?${repository.rootUri.fsPath}#${diffId}${DiffView.index++}`);
   }
 }