+++ src/views/processView.ts
@@ -6,6 +6,7 @@ import { processLog } from '../extension';
 import { View } from './general/view';
 import { MagitProcessLogEntry } from '../models/magitProcessLogEntry';
 import { TextView } from './general/textView';
+import { MagitState } from '../models/magitState';
 
 class ProcessLogEntryView extends View {
   isFoldable = true;
@@ -41,7 +42,7 @@ export default class ProcessView extends DocumentView {
     }
   }
 
-  public update(repository: MagitRepository): void {
+  public update(state: MagitState): void {
     this.provideContent();
     this.triggerUpdate();
   }